+++ carma_ros2_utils/src/timers/testing/TestTimerFactory.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020-2021 LEIDOS.
+ * Copyright (C) 2020-2022 LEIDOS.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-#include <carma_ros2_utils/timers/testing/TestTimerFactory.h>
+#include <carma_ros2_utils/timers/testing/TestTimerFactory.hpp>
 
 namespace carma_ros2_utils
 {
